+++ greenlight/src/greenlight/github_types.py
@@ -1,0 +1,128 @@
+"""Structural (Protocol) contracts the GitHub client speaks."""
+
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, Protocol
+
+if TYPE_CHECKING:
+    from collections.abc import Iterable
+    from datetime import datetime
+
+    class _PRUser(Protocol):
+        @property
+        def login(self) -> str | None: ...
+
+    class _Label(Protocol):
+        @property
+        def name(self) -> str: ...
+
+    class _PullRequest(Protocol):
+        @property
+        def number(self) -> int: ...
+        @property
+        def user(self) -> _PRUser | None: ...
+        @property
+        def title(self) -> str: ...
+        @property
+        def html_url(self) -> str: ...
+        @property
+        def head(self) -> _PRBase: ...
+        @property
+        def updated_at(self) -> datetime | None: ...
+        @property
+        def labels(self) -> Iterable[_Label]: ...
+        @property
+        def draft(self) -> bool: ...
+
+    class _Repo(Protocol):
+        def get_pulls(self, state: str) -> Iterable[_PullRequest]: ...
+
+    class _RepoClient(Protocol):
+        def get_repo(self, full_name_or_id: str) -> _Repo: ...
+
+    class _PRActor(Protocol):
+        @property
+        def login(self) -> str | None: ...
+        @property
+        def type(self) -> str: ...
+
+    class _PRComment(Protocol):
+        @property
+        def id(self) -> int: ...
+        @property
+        def user(self) -> _PRActor | None: ...
+        @property
+        def body(self) -> str: ...
+
+    class _PRReview(Protocol):
+        @property
+        def id(self) -> int: ...
+        @property
+        def user(self) -> _PRActor | None: ...
+        @property
+        def body(self) -> str: ...
+        @property
+        def state(self) -> str: ...
+
+    class _PRBase(Protocol):
+        @property
+        def sha(self) -> str: ...
+
+    class _FingerprintPR(Protocol):
+        @property
+        def head(self) -> _PRBase: ...
+        def get_issue_comments(self) -> Iterable[_PRComment]: ...
+        def get_review_comments(self) -> Iterable[_PRComment]: ...
+        def get_reviews(self) -> Iterable[_PRReview]: ...
+
+    class _ScanRepo(Protocol):
+        def get_pull(self, number: int) -> _FingerprintPR: ...
+
+    class _AuthorPR(Protocol):
+        @property
+        def user(self) -> _PRUser | None: ...
+
+    class _AuthorRepo(Protocol):
+        def get_pull(self, number: int) -> _AuthorPR: ...
+
+    class _AuthorClient(Protocol):
+        def get_repo(self, full_name_or_id: str) -> _AuthorRepo: ...
+
+    class _VerdictReview(Protocol):
+        @property
+        def id(self) -> int: ...
+        @property
+        def user(self) -> _PRUser | None: ...
+        @property
+        def state(self) -> str: ...
+        def dismiss(self, message: str) -> None: ...
+
+    class _VerdictComment(Protocol):
+        @property
+        def body(self) -> str: ...
+        @property
+        def user(self) -> _PRUser | None: ...
+        def edit(self, body: str) -> None: ...
+
+    class VerdictPR(Protocol):
+        @property
+        def head(self) -> _PRBase: ...
+        def create_review(self, *, body: str, event: str) -> object: ...
+        def create_issue_comment(self, body: str) -> object: ...
+        def get_issue_comments(self) -> Iterable[_VerdictComment]: ...
+        def get_reviews(self) -> Iterable[_VerdictReview]: ...
+
+    class _VerdictRepo(Protocol):
+        def get_pull(self, number: int) -> VerdictPR: ...
+
+
+class VerdictClient(Protocol):
+    """Structural GitHub client for the verdict path; the real ``github.Github`` satisfies it."""
+
+    def get_repo(self, full_name_or_id: str) -> _VerdictRepo: ...
+
+
+class ScanClient(Protocol):
+    """Structural GitHub client for the scan/fingerprint path; the real ``github.Github`` satisfies it."""
+
+    def get_repo(self, full_name_or_id: str) -> _ScanRepo: ...
